+++ coachcoach/common/src/main/java/com/coachcoach/common/api/InsightQueryApi.java
@@ -4,5 +4,5 @@ import java.time.LocalDate;
 
 public interface InsightQueryApi {
     void deleteByUserId(Long userId);
-    int getNumOfDangerMenus(Long userId);
+    Long getNumOfDangerMenus(Long userId);
 }
